Refactor the function that creates the grid.
Change the function that creates the grid so that
rows can be appended to columns and thus it can be
styled with flex. Styled appropriately with flex
in such a way that the rows and column cells
will always shrink or grow to fit in evenly in their
container and never change container size.