Refactor the newUserGrid() function
The IF statement checks if the value is
a number, not less than one and not
greater than 100. if the conditions are
satisfied, it clears the old grid and
creates a new one based off the value
of the prompt. Also changed the
button id in the HTML and effected
it in the javascript.