Create a rainbowOnMouseOver() function and refactor the eraseBackgroundColor() function
added a click event listener to the Rainbow button.
On click, it adds a mouseover event listener to
the grid columns (items) and change their background
color to the getRandomColor() function.

on the eraser button and on click, it adds a mouseover
event listener to the grid columns (items) and changes
their background color to white.