Create the increaseOpacity function and refactor the eraseBackgroundColor function
The function listens to the click event on the
increaseOpacityButton and on click, it adds a
mouseover event listener to the grid columns (items) and
gives them a class of increase-background-opacity.
Added this class to the css with the property ans value of
background: rgba(0,0,0,0.1). The function adds a background
color of black to the grid items. It then sets the opacity
to (parseFloat(target.style.opacity) || 0) + 0.1;

Also refactored the eraseBackgroundColor to
remove the increae-background-opacity class and
set the backgroundColor to transparent instead of white