Refactor the newUserGrid()
Refactor the new user grid to accept user
value via a range slider.On the index.html,
created a span close to the slider to display
the grid dimension. Added the newUserGrid()
to the createGrid() so that i can also show
the grid dimension on the on load even without
interacting with the slider.