Add a info icon (&#9432) on the Increasing Opacity Button
Also added a title attribute to the button also
in the index.html so that the user gets more
info about the button's feature on hover. Added
and a paragraph with class .instructions that
gives instructions about the buttons. Styled the
paragraph in the css with display of flex,
justified content and align items center.

Increased the 'blueness' of the buttons
before hover